requirements.txt
telethon
python-dotenv
pydantic-settings
pymongo
aiosqlite
motor